requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
openai==1.62.0
pydantic==2.10.6
